split world from main file
